create model, repository, service and controller of image and photo. Verify user delete post and user delete comment. Remove trace to exception response.

+++ src/main/java/com/framework/blog/service/CommentService.java
@@ -1,10 +1,15 @@
 package com.framework.blog.service;
 
 import com.framework.blog.exception.CommentNotExist;
+import com.framework.blog.exception.PermissionDeniedException;
 import com.framework.blog.model.Comment;
+import com.framework.blog.model.UserBlog;
 import com.framework.blog.repository.CommentRepository;
+import com.framework.blog.repository.UserBlogRepository;
+import org.springframework.dao.PermissionDeniedDataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import javax.validation.Valid;
@@ -13,17 +18,27 @@ import javax.validation.Valid;
 public class CommentService {
 
     private final CommentRepository commentRepository;
+    private final UserBlogRepository userBlogRepository;
 
-    public CommentService(CommentRepository commentRepository) {
+    public CommentService(CommentRepository commentRepository, UserBlogRepository userBlogRepository) {
         this.commentRepository = commentRepository;
+        this.userBlogRepository = userBlogRepository;
     }
 
-    public Comment createComment(@Valid Comment comment){
+    public Comment createComment(Comment comment){
+        UserBlog userLogged = userLogged();
+        comment.setUserBlog(userLogged);
         return commentRepository.save(comment);
     }
 
-    public void deleteComment(Long id) throws CommentNotExist {
-        commentRepository.deleteById(checkCommentExist(id).getId());
+    public void deleteComment(Long id) throws CommentNotExist, PermissionDeniedException {
+        UserBlog userLogged = userLogged();
+        Comment comment = checkCommentExist(id);
+        if (comment.getUserBlog().getId() == userLogged.getId()){
+            commentRepository.deleteById(checkCommentExist(id).getId());
+        }else{
+            throw new PermissionDeniedException("You don't have permission to delete this comment!");
+        }
     }
 
     public Comment findById(Long id) throws CommentNotExist {
@@ -46,5 +61,10 @@ public class CommentService {
         return commentRepository.findById(id)
                 .orElseThrow(()-> new CommentNotExist("Cannot find comment with id: "+ id +" /n" +
                         "Please insert another id and try again!"));
+    }
+
+    private UserBlog userLogged() {
+        String username = (String) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        return userBlogRepository.findByUsername(username).get();
     }
 }
